ChatProtocolAgent/google_scholar_agent/requirements.txt
uagents==0.22.3
google-generativeai==0.8.5
python-dotenv==1.0.1
uagents_core==0.2.3
scholarly==1.7.11
